backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pytest
python-multipart
web3
tensorflow
numpy
pillow
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
